+++ engine/include/map.h
@@ -2,7 +2,7 @@
 #define _MSX_H_MAP
 
 
-void map_inflate(const byte * dict, const byte * in, byte * out, uint data_size,
+void map_inflate(const byte * dict, const byte * in, byte * out, uint16_t data_size,
 		 byte w);
 void map_inflate_screen(const byte * dict, const byte * in, byte * out, byte w, byte vpx, byte vpy);
